feat: 결제 엔티티 및 포트원 설정 추가
- Payment 엔티티 추가 (PaymentStatus, PgProvider, PaymentMethod 포함)
- 결제 관련 DTO 추가 (PreparePaymentRequest/Response, CompletePaymentRequest/Response)
- 포트원 설정 추가 (application.yml, PortOneProperties, PortOneClientConfig)
- Course.estimatedCost, Reservation.quotedAmount Integer → Long 통일

+++ apps/be/src/main/java/com/breadbread/course/dto/CourseSummaryResponse.java
@@ -14,7 +14,7 @@ public class CourseSummaryResponse {
     private String thumbnailUrl;
     private int bakeryCount;
     private String estimatedTime;
-    private Integer estimatedCost;
+    private Long estimatedCost;
     private int likeCount;
     private boolean liked;
     private List<CourseBakerySummary> bakeries;
