+++ apps/be/src/main/java/com/breadbread/course/dto/UpdateCourseRequest.java
@@ -18,7 +18,7 @@ public class UpdateCourseRequest {
     private String estimatedTime;
 
     @Min(0)
-    private Integer estimatedCost;
+    private Long estimatedCost;
 
     // ManualCourseInfo
     private Boolean editorPick;
